requirements.txt
acachecontrol==0.3.5
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
click==8.1.7
colorama==0.4.6
decorator==5.1.1
fastapi==0.115.3
fastapi-login==1.10.2
frozenlist==1.5.0
h11==0.14.0
idna==3.10
multidict==6.1.0
neo4j==5.25.0
okta-jwt-verifier==0.2.7
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
retry2==0.9.5
setuptools==75.3.0
sniffio==1.3.1
starlette==0.41.0
typing_extensions==4.12.2
uvicorn==0.32.0
yarl==1.17.0
